- Restructured project layout - Added setup.py to ensure .env variables exist and vector store exists
