+++ .gitignore
@@ -3,4 +3,4 @@ __pycache__/
 
 # User sensitive data
 fine_tuning/**/*.json*
-knowledgebase/**/*.json*
+knowledge_base/**/*.json*
