+++ knowledge_base/people/utils.py
@@ -1,7 +1,7 @@
 import json
 
 def get_people() -> list[dict]:
-    with open(f"knowledgebase/people/people.json", "r") as f:
+    with open(f"knowledge_base/people/people.json", "r") as f:
         return json.load(f)
 
 def get_user_info() -> dict:
@@ -9,7 +9,7 @@ def get_user_info() -> dict:
     for person in people:
         if "self" in person["relations"]:
             return person
-    print("WARNING: No user info found in knowledgebase/people/people.json!")
+    print("WARNING: No user info found in knowledge_base/people/people.json!")
     return None
 
 def get_person_info_by_sender_id(sender_id: str) -> dict:
