+++ tools/tools.py
@@ -1,22 +1,9 @@
-import glob
 from datetime import datetime
-from tools.utils import get_tools_list
-import knowledgebase.people.tools as people_tools
+import knowledge_base.people.tools as people_tools
 
 # https://platform.openai.com/docs/guides/function-calling
 # If your function has no return value (e.g. send_email),
 # simply return a string to indicate success or failure. (e.g. "success")
-
-# Create tools list
-tools_list = get_tools_list(__file__, avoid_functions=["call_function"], web_search=True, file_search=True)
-tools_files = glob.glob("knowledgebase/**/tools.py", recursive=True)
-for file_path in tools_files:
-    folder_name = file_path.split("/")[-2]
-    tools_list += get_tools_list(
-        file_path,
-        module_name=folder_name,
-        avoid_functions=people_tools.avoid_functions
-    )
 
 def get_current_date() -> str:
     """
@@ -35,10 +22,12 @@ def get_current_time() -> str:
     return datetime.now().strftime("%I:%M %p")
 
 def call_function(name: str, args: dict):
-    if name == "get_current_date":
-        return get_current_date(**args)
-    if name == "get_current_time":
-        return get_current_time(**args)
-    if name.startswith("people-"):
-        return people_tools.call_function(name.split("-")[1], args)
-    raise ValueError(f"Function '{name}' not found")
+    module, func = name.split("-")
+    if module == "tools":
+        if func == "get_current_date":
+            return get_current_date(**args)
+        if func == "get_current_time":
+            return get_current_time(**args)
+    if module == "people":
+        return people_tools.call_function(func, args)
+    raise ValueError(f"Function '{module}.{func}' not found")
